# api/swagger/swagger.yaml
basePath: /v1
definitions:
  authentication.LoginRequest:
    properties:
      password:
        maxLength: 100
        minLength: 8
        type: string
      username:
        maxLength: 100
        minLength: 2
        type: string
    required:
    - password
    - username
    type: object
  response.Response:
    properties:
      data: {}
      detail:
        type: string
      errors: {}
      title:
        type: string
    type: object
  tapmoney.InquiryRequest:
    properties:
      amount:
        maximum: 1000000
        minimum: 10000
        type: integer
      card_number:
        maxLength: 19
        minLength: 16
        type: string
      source_account:
        maxLength: 19
        minLength: 16
        type: string
    required:
    - amount
    - card_number
    - source_account
    type: object
  tapmoney.PaymentRequest:
    properties:
      amount:
        maximum: 1000000
        minimum: 10000
        type: integer
      notes:
        maxLength: 255
        type: string
      transaction_id:
        type: string
    required:
    - amount
    - transaction_id
    type: object
  transfer.InitiateRequest:
    properties:
      amount:
        maximum: 50000000
        minimum: 1000
        type: integer
      destination_account:
        type: string
      note:
        type: string
      source_account:
        type: string
    required:
    - amount
    - destination_account
    - source_account
    type: object
  transfer.ProcessRequest:
    properties:
      amount:
        type: integer
      destination_account:
        type: string
      source_account:
        type: string
      uuid:
        type: string
    required:
    - amount
    - destination_account
    - source_account
    - uuid
    type: object
host: api.bankkrud.com
info:
  contact:
    email: billyimmcul2010@gmail.com
    name: Billy Kore
    url: https://www.swagger.io/support
  description: Bankfrud service API specification.
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: https://swagger.io/terms/
  title: API Specification
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: User login
      parameters:
      - description: Login Request
        in: body
        name: LoginRequest
        required: true
        schema:
          $ref: '#/definitions/authentication.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: User login
      tags:
      - authentication
  /tapmoney/inquiry:
    post:
      consumes:
      - application/json
      description: TapMoney transaction inquiry process
      parameters:
      - description: Authorization token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Inquiry Request
        in: body
        name: InquiryRequest
        required: true
        schema:
          $ref: '#/definitions/tapmoney.InquiryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: TapMoney inquiry
      tags:
      - tapmoney
  /tapmoney/payment:
    post:
      consumes:
      - application/json
      description: TapMoney transaction payment process
      parameters:
      - description: Authorization token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Payment Request
        in: body
        name: PaymentRequest
        required: true
        schema:
          $ref: '#/definitions/tapmoney.PaymentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: TapMoney payment
      tags:
      - tapmoney
  /transactions:
    get:
      consumes:
      - application/json
      description: Get transactions with filters
      parameters:
      - description: Authorization token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Transaction UUID
        in: query
        name: uuid
        type: string
      - description: Transaction Type
        in: query
        name: transaction_type
        type: string
      - description: Source Account UUID
        in: query
        name: source_account
        type: string
      - description: Target Account UUID
        in: query
        name: target_account
        type: string
      - description: Transaction Status
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Get transactions
      tags:
      - transactions
  /transactions/{uuid}:
    get:
      consumes:
      - application/json
      description: Get transaction by UUID
      parameters:
      - description: Authorization token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Transaction UUID
        in: path
        name: uuid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Get transaction by UUID
      tags:
      - transactions
  /transfers/{uuid}:
    get:
      consumes:
      - application/json
      description: Get transfer detail by uuid
      parameters:
      - description: Authorization token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Transfer UUID
        in: path
        name: uuid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Get transfer detail
      tags:
      - transfers
  /transfers/{uuid}/process:
    post:
      consumes:
      - application/json
      description: Process transfer transaction
      parameters:
      - description: Authorization token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Process Transfer Request
        in: body
        name: ProcessRequest
        required: true
        schema:
          $ref: '#/definitions/transfer.ProcessRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Process transfer
      tags:
      - transfers
  /transfers/init:
    post:
      consumes:
      - application/json
      description: Initiate transfer transaction
      parameters:
      - description: Authorization token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Initiate Transfer Request
        in: body
        name: InitiateRequest
        required: true
        schema:
          $ref: '#/definitions/transfer.InitiateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Initiate transfer
      tags:
      - transfer
  /users/me:
    get:
      consumes:
      - application/json
      description: Get logged in user details
      parameters:
      - description: Authorization token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Fields
        in: query
        name: fields
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Get logged in user
      tags:
      - users
schemes:
- http
- https
swagger: "2.0"